+++ LibIO/StorageTypes.h
@@ -2,7 +2,9 @@
 #ifndef __STORAGE_TYPES_H
 #define __STORAGE_TYPES_H
 
-typedef unsigned long FileOffset;
+#include <stdint.h>
+
+typedef uint32_t FileOffset;
 
 // used internally for offsets inside files:
 const FileOffset INVALID_OFFSET = 0xffffffff;
